services/rakuten-renewal/terraform/sentinel/requirements.txt
functions-framework>=3.5
google-cloud-firestore>=2.16
google-cloud-aiplatform>=1.48
sendgrid>=6.11
